hud now follows camera flawless

+++ game/button.cpp
@@ -30,15 +30,13 @@ Button::~Button() {
 }
 
 void Button::update(float deltaTime) {
-	int mousex = parent->input()->getMouseX() + parent->camera()->position.x - SWIDTH / 2;
-	int mousey = parent->input()->getMouseY() + parent->camera()->position.y - SHEIGHT / 2;
+	float mousex = parent->input()->getMouseX() + parent->camera()->position.x - SWIDTH / 2;
+	float mousey = parent->input()->getMouseY() + parent->camera()->position.y - SHEIGHT / 2;
 
-	int buttonLeft = pos.x - (size.x / 2);
-	int buttonRight = pos.x + (size.x / 2);
-	int buttonTop = pos.y + (size.y / 2);
-	int buttonDown = pos.y - (size.y / 2);
-
-	//std::cout << pos.x << std::endl;
+	float buttonLeft = pos.x - (size.x / 2);
+	float buttonRight = pos.x + (size.x / 2);
+	float buttonTop = pos.y + (size.y / 2);
+	float buttonDown = pos.y - (size.y / 2);
 
 	//If you hover...
 	if (mousex > buttonLeft && mousex < buttonRight && buttonTop > mousey && buttonDown < mousey) {
